feat: Milestone 추가 시 필요한 정보만 넘기도록 수정
Milestone 추가시에는 불필요한 부분(id, issue, ...)들을 넘겨줄 필요가 없다고 판단

+++ app/src/main/java/com/example/it/issuetracker/data/datasource/MilestoneFakeDatabase.kt
@@ -36,8 +36,8 @@ object MilestoneFakeDatabase {
         )
     )
 
-    val size: Int
-        get() = database.size
+    val size: Long
+        get() = database.size.toLong()
 
     suspend fun add(milestoneDto: MilestoneDto) {
         delay(1000)
